backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
Pillow>=10.0.0
python-dotenv>=1.0.0
whitenoise>=6.4.0
gunicorn>=20.1.0
psycopg2-binary>=2.9.5
django-cors-headers>=4.1.0
django-environ>=0.10.0
